requirements.txt
PySide6>=6.6.0
opencv-python>=4.8.0
numpy>=1.26.0
watchdog>=3.0.0
tifffile>=2023.8.30
Pillow>=10.0.0
ExifRead>=3.0.0
